refactor: extract GitHub PR comment helpers
